FIX: Warnings

+++ Examples/Numerics/FourierDescriptors1.cxx
@@ -175,7 +175,7 @@ int main(int argc, char * argv[] )
   //  Software Guide : EndLatex 
 
   // Software Guide : BeginCodeSnippet
-  const unsigned int powerOfTwo   = ceil( log( (double)(numberOfPoints)) /
+  const unsigned int powerOfTwo   = (unsigned int)ceil( log( (double)(numberOfPoints)) /
                                           log( (double)(2.0)));
 
   const unsigned int spectrumSize = 1 << powerOfTwo;
